+++ packages/styled-components/src/models/StyledNativeComponent.js
@@ -42,7 +42,7 @@ class StyledNativeComponent extends Component<*, *> {
           // eslint-disable-next-line no-console
           console.warn(
             `Functions as object-form attrs({}) keys are now deprecated and will be removed in a future version of styled-components. Switch to the new attrs(props => ({})) syntax instead for easier and more powerful composition. The attrs key in question is "${key}" on component "${displayName}".`,
-            `\n ${(new Error()).stack}`
+            `\n ${new Error().stack}`
           )
       );
 
@@ -65,6 +65,7 @@ class StyledNativeComponent extends Component<*, *> {
         {(theme?: Theme) => {
           const {
             as: renderAs,
+            forwardedAs,
             forwardedComponent,
             forwardedRef,
             innerRef,
@@ -74,13 +75,10 @@ class StyledNativeComponent extends Component<*, *> {
 
           const { defaultProps, displayName, target } = forwardedComponent;
 
-          let generatedStyles;
-          if (theme !== undefined) {
-            const themeProp = determineTheme(this.props, theme, defaultProps);
-            generatedStyles = this.generateAndInjectStyles(themeProp, this.props);
-          } else {
-            generatedStyles = this.generateAndInjectStyles(theme || EMPTY_OBJECT, this.props);
-          }
+          const generatedStyles = this.generateAndInjectStyles(
+            determineTheme(this.props, theme, defaultProps) || EMPTY_OBJECT,
+            this.props
+          );
 
           const propsForElement = {
             ...this.attrs,
@@ -88,6 +86,7 @@ class StyledNativeComponent extends Component<*, *> {
             style: [generatedStyles].concat(style),
           };
 
+          if (forwardedAs) propsForElement.as = forwardedAs;
           if (forwardedRef) propsForElement.ref = forwardedRef;
 
           if (process.env.NODE_ENV !== 'production' && innerRef) {